move DeleteManyAsync from UserService into DeleteManagedOrganizationUserAccountCommand
