+++ src/Core/AdminConsole/OrganizationFeatures/OrganizationUsers/DeleteManagedOrganizationUserAccountCommand.cs
@@ -1,10 +1,14 @@
 ﻿using Bit.Core.AdminConsole.OrganizationFeatures.OrganizationUsers.Interfaces;
+using Bit.Core.AdminConsole.Repositories;
 using Bit.Core.Context;
 using Bit.Core.Entities;
 using Bit.Core.Enums;
 using Bit.Core.Exceptions;
 using Bit.Core.Repositories;
 using Bit.Core.Services;
+using Bit.Core.Tools.Enums;
+using Bit.Core.Tools.Models.Business;
+using Bit.Core.Tools.Services;
 
 #nullable enable
 
@@ -20,7 +24,10 @@ public class DeleteManagedOrganizationUserAccountCommand : IDeleteManagedOrganiz
     private readonly ICurrentContext _currentContext;
     private readonly IHasConfirmedOwnersExceptQuery _hasConfirmedOwnersExceptQuery;
     private readonly IFeatureService _featureService;
-
+    private readonly IReferenceEventService _referenceEventService;
+    private readonly IPushNotificationService _pushService;
+    private readonly IOrganizationRepository _organizationRepository;
+    private readonly IProviderUserRepository _providerUserRepository;
     public DeleteManagedOrganizationUserAccountCommand(
         IUserService userService,
         IEventService eventService,
@@ -29,7 +36,11 @@ public class DeleteManagedOrganizationUserAccountCommand : IDeleteManagedOrganiz
         IUserRepository userRepository,
         ICurrentContext currentContext,
         IHasConfirmedOwnersExceptQuery hasConfirmedOwnersExceptQuery,
-        IFeatureService featureService)
+        IFeatureService featureService,
+        IReferenceEventService referenceEventService,
+        IPushNotificationService pushService,
+        IOrganizationRepository organizationRepository,
+        IProviderUserRepository providerUserRepository)
     {
         _userService = userService;
         _eventService = eventService;
@@ -39,6 +50,10 @@ public class DeleteManagedOrganizationUserAccountCommand : IDeleteManagedOrganiz
         _currentContext = currentContext;
         _hasConfirmedOwnersExceptQuery = hasConfirmedOwnersExceptQuery;
         _featureService = featureService;
+        _referenceEventService = referenceEventService;
+        _pushService = pushService;
+        _organizationRepository = organizationRepository;
+        _providerUserRepository = providerUserRepository;
     }
 
     public async Task DeleteUserAsync(Guid organizationId, Guid organizationUserId, Guid? deletingUserId)
@@ -99,7 +114,7 @@ public class DeleteManagedOrganizationUserAccountCommand : IDeleteManagedOrganiz
             }
         }
 
-        await _userService.DeleteManyAsync(users);
+        await DeleteManyAsync(users);
         await LogDeletedOrganizationUsersAsync(orgUsers, results);
 
         return results;
@@ -159,5 +174,75 @@ public class DeleteManagedOrganizationUserAccountCommand : IDeleteManagedOrganiz
         {
             await _eventService.LogOrganizationUserEventsAsync(events);
         }
+    }
+    private async Task<IEnumerable<(Guid UserId, string? ErrorMessage)>> DeleteManyAsync(IEnumerable<User> users)
+    {
+        var results = await ValidateDeleteUsersAsync(users.ToList());
+
+        var usersToDelete = users.Where(user => results.Select(r => r.UserId == user.Id && string.IsNullOrEmpty(r.ErrorMessage)).Count() > 1);
+
+        await _userRepository.DeleteManyAsync(usersToDelete);
+        foreach (var user in usersToDelete)
+        {
+            await _referenceEventService.RaiseEventAsync(
+                new ReferenceEvent(ReferenceEventType.DeleteAccount, user, _currentContext));
+            await _pushService.PushLogOutAsync(user.Id);
+        }
+
+        return results;
+    }
+
+    private async Task<IEnumerable<(Guid UserId, string? ErrorMessage)>> ValidateDeleteUsersAsync(List<User> users)
+    {
+        var userResult = new List<(Guid UserId, string? ErrorMessage)>();
+
+        foreach (var user in users)
+        {
+            // Check if user is the only owner of any organizations.
+            var onlyOwnerCount = await _organizationUserRepository.GetCountByOnlyOwnerAsync(user.Id);
+            if (onlyOwnerCount > 0)
+            {
+                userResult.Add((user.Id, "Cannot delete this user because it is the sole owner of at least one organization. Please delete these organizations or upgrade another user."));
+                continue;
+            }
+
+            var orgs = await _organizationUserRepository.GetManyDetailsByUserAsync(user.Id, OrganizationUserStatusType.Confirmed);
+            if (orgs.Count == 1)
+            {
+                var org = await _organizationRepository.GetByIdAsync(orgs.First().OrganizationId);
+                if (org != null && (!org.Enabled || string.IsNullOrWhiteSpace(org.GatewaySubscriptionId)))
+                {
+                    var orgCount = await _organizationUserRepository.GetCountByOrganizationIdAsync(org.Id);
+                    if (orgCount <= 1)
+                    {
+                        await _organizationRepository.DeleteAsync(org);
+                    }
+                    else
+                    {
+                        userResult.Add((user.Id, "Cannot delete this user because it is the sole owner of at least one organization. Please delete these organizations or upgrade another user."));
+                        continue;
+                    }
+                }
+            }
+
+            var onlyOwnerProviderCount = await _providerUserRepository.GetCountByOnlyOwnerAsync(user.Id);
+            if (onlyOwnerProviderCount > 0)
+            {
+                userResult.Add((user.Id, "Cannot delete this user because it is the sole owner of at least one provider. Please delete these providers or upgrade another user."));
+                continue;
+            }
+
+            if (!string.IsNullOrWhiteSpace(user.GatewaySubscriptionId))
+            {
+                try
+                {
+                    await _userService.CancelPremiumAsync(user);
+                }
+                catch (GatewayException) { }
+            }
+
+            userResult.Add((user.Id, string.Empty));
+        }
+        return userResult;
     }
 }
