+++ test/Core.Test/Services/UserServiceTests.cs
@@ -379,40 +379,6 @@ public class UserServiceTests
         Assert.True(result == IdentityResult.Success);
     }
 
-    [Theory, BitAutoData]
-    public async Task DeleteManyAsync_WithAccountDeprovisioningEnabled_Succeeds(
-        SutProvider<UserService> sutProvider, User user1, User user2, User user3
-    )
-    {
-        // Arrange
-        var userList = new List<User> { user1, user2 };
-        sutProvider.GetDependency<IFeatureService>().IsEnabled(FeatureFlagKeys.AccountDeprovisioning).Returns(true);
-
-        // Act
-        var results = await sutProvider.Sut.DeleteManyAsync(userList);
-
-        // Assert
-        Assert.True(results.Where(result => result.UserId == user1.Id || result.UserId == user2.Id).Count() == 2);
-        Assert.True(results.Where(result => result.UserId == user3.Id).Count() == 0);
-    }
-
-    [Theory, BitAutoData]
-    public async Task DeleteManyAsync_WithAccountDeprovisioningDisabled_Succeeds(
-        SutProvider<UserService> sutProvider, User user1, User user2, User user3
-    )
-    {
-        // Arrange
-        var userList = new List<User> { user1, user2 };
-        sutProvider.GetDependency<IFeatureService>().IsEnabled(FeatureFlagKeys.AccountDeprovisioning).Returns(false);
-
-        // Act
-        var results = await sutProvider.Sut.DeleteManyAsync(userList);
-
-        // Assert
-        Assert.True(results.Where(result => result.UserId == user1.Id || result.UserId == user2.Id).Count() == 2);
-        Assert.True(results.Where(result => result.UserId == user3.Id).Count() == 0);
-    }
-
     private static void SetupUserAndDevice(User user,
         bool shouldHavePassword)
     {
